+++ src/iterators/async_iterators/mod.rs
@@ -3,11 +3,15 @@
 
 //! Async iterators.
 
+use crate::MRBIterator;
+use crate::Storage;
 use crate::iterators::async_iterators::detached::AsyncDetached;
 use crate::iterators::util_macros::delegate;
 use crate::iterators::util_macros::muncher;
-use crate::{MRBIterator, MutRB};
+use crate::ring_buffer::variants::async_rb::AsyncMutRingBuf;
+use crate::ring_buffer::wrappers::buf_ref::BufRef;
 use core::future::Future;
+use core::marker::PhantomData;
 use core::pin::Pin;
 use core::task::{Context, Poll, Waker};
 
@@ -17,20 +21,22 @@ pub(crate) mod prod_iter;
 pub(crate) mod work_iter;
 
 /// Trait implemented by async iterators.
-pub trait AsyncIterator {
+pub trait AsyncIterator<'buf> {
     type I: MRBIterator;
-    type B: MutRB;
+    type S: Storage;
 
-    fn register_waker(&mut self, waker: &Waker);
+    fn register_waker(&self, waker: &Waker);
+    fn take_waker(&self) -> Option<Waker>;
+    fn wake_next(&self);
 
     fn inner(&self) -> &Self::I;
     fn inner_mut(&mut self) -> &mut Self::I;
 
     fn into_sync(self) -> Self::I;
 
-    fn from_sync(iter: Self::I) -> Self;
+    fn from_sync(iter: Self::I, waker: BufRef<'buf, AsyncMutRingBuf<Self::S>>) -> Self;
 
-    fn detach(self) -> AsyncDetached<Self, Self::B>
+    fn detach(self) -> AsyncDetached<'buf, Self, AsyncMutRingBuf<Self::S>>
     where
         Self: Sized,
     {
@@ -45,75 +51,65 @@ pub trait AsyncIterator {
     delegate!(MRBIterator, fn cons_index(&self) -> usize);
     delegate!(MRBIterator, fn index(&self) -> usize);
     delegate!(MRBIterator, fn available(&(mut) self) -> usize);
-    delegate!(MRBIterator, unsafe fn advance(&(mut) self, count: usize));
 }
 
 /// Future returned by methods in async iterators.
-pub struct MRBFuture<'a, I, P, O, const R: bool>
+pub struct MRBFuture<'buf, 'a, I, P, O, const R: bool>
 where
-    I: AsyncIterator,
+    I: AsyncIterator<'buf>,
 {
     iter: &'a mut I,
     p: Option<P>,
     f_r: Option<fn(&mut I, &mut P) -> Option<O>>,
     f_m: Option<fn(&mut I, P) -> Result<O, P>>,
+    phantom: PhantomData<&'buf ()>,
 }
 
-impl<I: AsyncIterator, P, O, const R: bool> Unpin for MRBFuture<'_, I, P, O, R> {}
+impl<'buf, 'a, I: AsyncIterator<'buf>, P, O, const R: bool> Unpin
+    for MRBFuture<'buf, 'a, I, P, O, R>
+{
+}
 
-impl<I: AsyncIterator, P, O, const R: bool> Future for MRBFuture<'_, I, P, O, R> {
+impl<'buf, 'a, I: AsyncIterator<'buf>, P, O, const R: bool> Future
+    for MRBFuture<'buf, 'a, I, P, O, R>
+{
     type Output = Option<O>;
 
     fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        let mut waker_registered = false;
         let f_r = self.f_r.take();
         let f_m = self.f_m.take();
+        let p = self.p.take();
 
-        loop {
-            let res = if R {
-                let mut p = self.p.take().unwrap();
-                let ret = f_r.as_ref().unwrap()(self.iter, &mut p);
-                ret.ok_or(p)
-            } else {
-                let p = self.p.take().unwrap();
-                f_m.as_ref().unwrap()(self.iter, p)
-            };
+        let res = if R {
+            let mut p = p.unwrap();
+            let ret = f_r.as_ref().unwrap()(self.iter, &mut p);
+            ret.ok_or(p)
+        } else {
+            f_m.as_ref().unwrap()(self.iter, p.unwrap())
+        };
 
-            match res {
-                Ok(r) => {
-                    break Poll::Ready(Some(r));
-                }
-                Err(p) => {
-                    self.f_r = f_r;
-                    self.f_m = f_m;
-                    self.p = Some(p);
-                }
+        match res {
+            Ok(r) => {
+                return Poll::Ready(Some(r));
             }
-
-            if waker_registered {
-                break Poll::Pending;
+            Err(p) => {
+                self.f_r = f_r;
+                self.f_m = f_m;
+                self.p = Some(p);
             }
-            self.iter.register_waker(cx.waker());
-            waker_registered = true;
         }
+
+        self.iter.register_waker(cx.waker());
+        Poll::Pending
     }
 }
 
 pub(crate) mod async_macros {
-
-    macro_rules! waker_registerer {
-        () => {
-            fn register_waker(&mut self, waker: &Waker) {
-                self.waker.get_or_insert(waker.clone()).clone_from(waker);
-            }
-        };
-    }
-
     macro_rules! gen_common_futs_fn {
-        ($(($CT: ty))*) => {
+        ($LT: lifetime) => {
             /// Async version of [`MRBIterator::get_workable`].
-            pub fn get_workable<'b>(&'_ mut self) -> MRBFuture<'_, Self, (), &'b mut T, true> {
-                fn f<'b, II: MRBIterator<Item = T>, I: AsyncIterator<I = II>, T>(
+            pub fn get_workable<'b>(&'b mut self) -> MRBFuture<$LT, 'b, Self, (), &'b mut T, true> {
+                fn f<'buf, 'b, II: MRBIterator<Item = T>, I: AsyncIterator<'buf, I = II>, T>(
                     s: &mut I,
                     _: &mut (),
                 ) -> Option<&'b mut T> {
@@ -125,15 +121,16 @@ pub(crate) mod async_macros {
                     p: Some(()),
                     f_r: Some(f),
                     f_m: None,
+                    phantom: PhantomData,
                 }
             }
 
             /// Async version of [`MRBIterator::get_workable_slice_exact`].
             pub fn get_workable_slice_exact<'b>(
-                &'_ mut self,
+                &'b mut self,
                 count: usize,
-            ) -> MRBFuture<'_, Self, usize, MutableSlice<'b, T>, true> {
-                fn f<'b, II: MRBIterator<Item = T>, I: AsyncIterator<I = II>, T>(
+            ) -> MRBFuture<$LT, 'b, Self, usize, MutableSlice<'b, T>, true> {
+                fn f<'buf, 'b, II: MRBIterator<Item = T>, I: AsyncIterator<'buf, I = II>, T>(
                     s: &mut I,
                     count: &mut usize,
                 ) -> Option<MutableSlice<'b, T>> {
@@ -145,14 +142,15 @@ pub(crate) mod async_macros {
                     p: Some(count),
                     f_r: Some(f),
                     f_m: None,
+                    phantom: PhantomData,
                 }
             }
 
             /// Async version of [`MRBIterator::get_workable_slice_avail`].
             pub fn get_workable_slice_avail<'b>(
-                &'_ mut self,
-            ) -> MRBFuture<'_, Self, (), MutableSlice<'b, T>, true> {
-                fn f<'b, II: MRBIterator<Item = T>, I: AsyncIterator<I = II>, T>(
+                &'b mut self,
+            ) -> MRBFuture<$LT, 'b, Self, (), MutableSlice<'b, T>, true> {
+                fn f<'buf, 'b, II: MRBIterator<Item = T>, I: AsyncIterator<'buf, I = II>, T>(
                     s: &mut I,
                     _: &mut (),
                 ) -> Option<MutableSlice<'b, T>> {
@@ -164,15 +162,16 @@ pub(crate) mod async_macros {
                     p: Some(()),
                     f_r: Some(f),
                     f_m: None,
+                    phantom: PhantomData,
                 }
             }
 
             /// Async version of [`MRBIterator::get_workable_slice_multiple_of`].
             pub fn get_workable_slice_multiple_of<'b>(
-                &'_ mut self,
+                &'b mut self,
                 count: usize,
-            ) -> MRBFuture<'_, Self, usize, MutableSlice<'b, T>, true> {
-                fn f<'b, II: MRBIterator<Item = T>, I: AsyncIterator<I = II>, T>(
+            ) -> MRBFuture<$LT, 'b, Self, usize, MutableSlice<'b, T>, true> {
+                fn f<'buf, 'b, II: MRBIterator<Item = T>, I: AsyncIterator<'buf, I = II>, T>(
                     s: &mut I,
                     count: &mut usize,
                 ) -> Option<MutableSlice<'b, T>> {
@@ -184,10 +183,18 @@ pub(crate) mod async_macros {
                     p: Some(count),
                     f_r: Some(f),
                     f_m: None,
+                    phantom: PhantomData,
                 }
+            }
+
+            pub unsafe fn advance(&mut self, count: usize) {
+                unsafe {
+                    self.inner.advance(count);
+                }
+                self.wake_next();
             }
         };
     }
 
-    pub(crate) use {gen_common_futs_fn, waker_registerer};
+    pub(crate) use gen_common_futs_fn;
 }