+++ src/ring_buffer/variants/concurrent_rb.rs
@@ -3,10 +3,6 @@ use core::num::NonZeroUsize;
 use core::sync::atomic::Ordering::{Acquire, Release};
 use core::sync::atomic::{AtomicBool, AtomicUsize};
 
-#[cfg(any(feature = "async", doc))]
-use crate::iterators::{
-    AsyncConsIter, AsyncProdIter, AsyncWorkIter, async_iterators::AsyncIterator,
-};
 use crate::iterators::{ConsIter, ProdIter, WorkIter};
 use crate::ring_buffer::storage::Storage;
 use crate::ring_buffer::variants::ring_buffer_trait::{
@@ -47,94 +43,6 @@ impl<S: Storage<Item = T>, T> ConcurrentRB for ConcurrentMutRingBuf<S> {}
 impl_splits!(ConcurrentMutRingBuf);
 
 impl<S: Storage<Item = T>, T> ConcurrentMutRingBuf<S> {
-    /// Consumes the buffer, yielding three async iterators. See:
-    /// - [`AsyncProdIter`];
-    /// - [`AsyncWorkIter`];
-    /// - [`AsyncConsIter`].
-    #[cfg(all(feature = "alloc", any(feature = "async", doc)))]
-    pub fn split_mut_async<'buf>(
-        self,
-    ) -> (
-        AsyncProdIter<'buf, ConcurrentMutRingBuf<S>>,
-        AsyncWorkIter<'buf, ConcurrentMutRingBuf<S>>,
-        AsyncConsIter<'buf, ConcurrentMutRingBuf<S>, true>,
-    ) {
-        self.set_prod_alive(true);
-        self.set_work_alive(true);
-        self.set_cons_alive(true);
-
-        let r = BufRef::new(self);
-        (
-            AsyncProdIter::from_sync(ProdIter::new(r.clone())),
-            AsyncWorkIter::from_sync(WorkIter::new(r.clone())),
-            AsyncConsIter::from_sync(ConsIter::new(r)),
-        )
-    }
-
-    /// Consumes the buffer, yielding two async iterators. See:
-    /// - [`AsyncProdIter`];
-    /// - [`AsyncConsIter`].
-    #[cfg(all(not(feature = "alloc"), any(feature = "async", doc)))]
-    pub fn split_async(
-        &mut self,
-    ) -> (
-        AsyncProdIter<ConcurrentMutRingBuf<S>>,
-        AsyncConsIter<ConcurrentMutRingBuf<S>, false>,
-    ) {
-        self.set_prod_alive(true);
-        self.set_cons_alive(true);
-
-        let r = BufRef::from_ref(self);
-        (
-            AsyncProdIter::from_sync(ProdIter::new(r.clone())),
-            AsyncConsIter::from_sync(ConsIter::new(r)),
-        )
-    }
-
-    /// Consumes the buffer, yielding three async iterators. See:
-    /// - [`AsyncProdIter`];
-    /// - [`AsyncWorkIter`];
-    /// - [`AsyncConsIter`].
-    #[cfg(all(not(feature = "alloc"), any(feature = "async", doc)))]
-    pub fn split_mut_async(
-        &mut self,
-    ) -> (
-        AsyncProdIter<ConcurrentMutRingBuf<S>>,
-        AsyncWorkIter<ConcurrentMutRingBuf<S>>,
-        AsyncConsIter<ConcurrentMutRingBuf<S>, true>,
-    ) {
-        self.set_prod_alive(true);
-        self.set_work_alive(true);
-        self.set_cons_alive(true);
-
-        let r = BufRef::from_ref(self);
-        (
-            AsyncProdIter::from_sync(ProdIter::new(r.clone())),
-            AsyncWorkIter::from_sync(WorkIter::new(r.clone())),
-            AsyncConsIter::from_sync(ConsIter::new(r)),
-        )
-    }
-
-    /// Consumes the buffer, yielding two async iterators. See:
-    /// - [`AsyncProdIter`];
-    /// - [`AsyncConsIter`].
-    #[cfg(all(feature = "alloc", any(feature = "async", doc)))]
-    pub fn split_async<'buf>(
-        self,
-    ) -> (
-        AsyncProdIter<'buf, ConcurrentMutRingBuf<S>>,
-        AsyncConsIter<'buf, ConcurrentMutRingBuf<S>, false>,
-    ) {
-        self.set_prod_alive(true);
-        self.set_cons_alive(true);
-
-        let r = BufRef::new(self);
-        (
-            AsyncProdIter::from_sync(ProdIter::new(r.clone())),
-            AsyncConsIter::from_sync(ConsIter::new(r)),
-        )
-    }
-
     pub(crate) fn _from(value: S) -> ConcurrentMutRingBuf<S> {
         assert!(value.len() > 0);