+++ src/ring_buffer/variants/mod.rs
@@ -1,3 +1,4 @@
+pub mod async_rb;
 pub mod concurrent_rb;
 pub mod local_rb;
 pub mod ring_buffer_trait;